Function coreaudio_sys::AudioComponentRegister

pub unsafe extern "C" fn AudioComponentRegister(
    inDesc: *const AudioComponentDescription, 
    inName: CFStringRef, 
    inVersion: UInt32, 
    inFactory: AudioComponentFactoryFunction
) -> AudioComponent

@function AudioComponentRegister @abstract Dynamically registers an AudioComponent within the current process @discussion AudioComponents are registered either when found in appropriate bundles in the filesystem, or via this call. AudioComponents registered via this call are available only within the current process.

@param inDesc The AudioComponentDescription that describes the AudioComponent. Note that the registrar needs to be sure to set the flag kAudioComponentFlag_SandboxSafe in the componentFlags field of the AudioComponentDescription to indicate that the AudioComponent can be loaded directly into a sandboxed process. @param inName the AudioComponent's name @param inVersion the AudioComponent's version @param inFactory an AudioComponentFactoryFunction which will create instances of your AudioComponent @result an AudioComponent object