Function coreaudio_sys::AudioDeviceTranslateTime

pub unsafe extern "C" fn AudioDeviceTranslateTime(
    inDevice: AudioObjectID, 
    inTime: *const AudioTimeStamp, 
    outTime: *mut AudioTimeStamp
) -> OSStatus

@function AudioDeviceTranslateTime @abstract Translates the time in the AudioDevice's time base from one representation to another. Note that the device has to be running @param inDevice The AudioDevice whose time base governs the translation. @param inTime An AudioTimeStamp containing the time to be translated. @param outTime An AudioTimeStamp into which the translated time is put. On entry, the mFlags field specifies which representations to translate the input time into. Because not every device supports all time representations, on exit, the mFlags field will indicate which translations were actually done. @result An OSStatus indicating success or failure. kAudioHardwareNotRunningError will be returned if the AudioDevice isn't running.