Function coreaudio_sys::AudioObjectSetPropertyData

pub unsafe extern "C" fn AudioObjectSetPropertyData(
    inObjectID: AudioObjectID, 
    inAddress: *const AudioObjectPropertyAddress, 
    inQualifierDataSize: UInt32, 
    inQualifierData: *const c_void, 
    inDataSize: UInt32, 
    inData: *const c_void
) -> OSStatus

@function AudioObjectSetPropertyData @abstract Tells an AudioObject to change the value of the given property using the provided data. @discussion Note that the value of the property should not be considered changed until the HAL has called the listeners as many properties values are changed asynchronously. @param inObjectID The AudioObject to change. @param inAddress An AudioObjectPropertyAddress indicating which property is being changed. @param inQualifierDataSize A UInt32 indicating the size of the buffer pointed to by inQualifierData. Note that not all properties require qualification, in which case this value will be 0. @param inQualifierData, A buffer of data to be used in determining the data of the property being queried. Note that not all properties require qualification, in which case this value will be NULL. @param inDataSize A UInt32 indicating the size of the buffer pointed to by inData. @param inData The buffer containing the data to be used to change the property's value. @result An OSStatus indicating success or failure.