Function coreaudio_sys::AudioUnitGetPropertyInfo

pub unsafe extern "C" fn AudioUnitGetPropertyInfo(
    inUnit: AudioUnit, 
    inID: AudioUnitPropertyID, 
    inScope: AudioUnitScope, 
    inElement: AudioUnitElement, 
    outDataSize: *mut UInt32, 
    outWritable: *mut Boolean
) -> OSStatus

@function AudioUnitGetPropertyInfo @abstract retrieves information about a specified property @discussion The API can be used to retrieve both the size of the property, and whether it is writable or not. In order to get a general answer on the capability of an audio unit, this function should be called before the audio unit is initialized (as some properties are writable when the audio unit is initialized, and others not)

@param inUnit the audio unit @param inID the property identifier @param inScope the scope of the property @param inElement the element of the scope @param outDataSize if not null, then will retrieve the maximum size for the property. if null, then it is ignored @param outWritable if not null, then will retrieve whether the property can be written or not. if null, then it is ignored

@result noErr, or various audio unit errors related to properties