Function coreaudio_sys::AudioUnitRender

pub unsafe extern "C" fn AudioUnitRender(
    inUnit: AudioUnit, 
    ioActionFlags: *mut AudioUnitRenderActionFlags, 
    inTimeStamp: *const AudioTimeStamp, 
    inOutputBusNumber: UInt32, 
    inNumberFrames: UInt32, 
    ioData: *mut AudioBufferList
) -> OSStatus

@function AudioUnitRender @abstract the render operation where ioData will contain the results of the audio unit's render operations @discussion an audio unit will render the amount of audio data described by inNumberOfFrames and the results of that render will be contained within ioData. The caller should provide audio time stamps where at least the sample time is valid and it is incrementing sequentially from its previous call (so, the next time stamp will be the current time stamp + inNumberFrames) If the sample time is not incrementing sequentially, the audio unit will infer that there is some discontinuity with the timeline it is rendering for

The caller must provide a valid ioData AudioBufferList that matches the expected topology for the current audio format for the given bus. The buffer list can be of two variants: (1) If the mData pointers are non-null then the audio unit will render its output into those buffers. These buffers should be aligned to 16 byte boundaries (which is normally what malloc will return). (2) If the mData pointers are null, then the audio unit can provide pointers to its own buffers. In this case the audio unit is required to keep those buffers valid for the duration of the calling thread's I/O cycle

@param inUnit the audio unit @param ioActionFlags any appropriate action flags for the render operation @param inTimeStamp the time stamp that applies to this particular render operation. when rendering for multiple output buses the time stamp will generally be the same for each output bus, so the audio unit is able to determine without doubt that this the same render operation @param inOutputBusNumber the output bus to render for @param inNumberFrames the number of sample frames to render @param ioData the audio buffer list that the audio unit is to render into.

@result noErr, or an audio unit render error