Function coreaudio_sys::CFArrayCreate

pub unsafe extern "C" fn CFArrayCreate(
    allocator: CFAllocatorRef, 
    values: *mut *const c_void, 
    numValues: CFIndex, 
    callBacks: *const CFArrayCallBacks
) -> CFArrayRef

@function CFArrayCreate Creates a new immutable array with the given values. @param allocator The CFAllocator which should be used to allocate memory for the array and its storage for values. This parameter may be NULL in which case the current default CFAllocator is used. If this reference is not a valid CFAllocator, the behavior is undefined. @param values A C array of the pointer-sized values to be in the array. The values in the array are ordered in the same order in which they appear in this C array. This parameter may be NULL if the numValues parameter is 0. This C array is not changed or freed by this function. If this parameter is not a valid pointer to a C array of at least numValues pointers, the behavior is undefined. @param numValues The number of values to copy from the values C array into the CFArray. This number will be the count of the array. If this parameter is negative, or greater than the number of values actually in the value's C array, the behavior is undefined. @param callBacks A pointer to a CFArrayCallBacks structure initialized with the callbacks for the array to use on each value in the array. The retain callback will be used within this function, for example, to retain all of the new values from the values C array. A copy of the contents of the callbacks structure is made, so that a pointer to a structure on the stack can be passed in, or can be reused for multiple array creations. If the version field of this callbacks structure is not one of the defined ones for CFArray, the behavior is undefined. The retain field may be NULL, in which case the CFArray will do nothing to add a retain to the contained values for the array. The release field may be NULL, in which case the CFArray will do nothing to remove the array's retain (if any) on the values when the array is destroyed. If the copyDescription field is NULL, the array will create a simple description for the value. If the equal field is NULL, the array will use pointer equality to test for equality of values. This callbacks parameter itself may be NULL, which is treated as if a valid structure of version 0 with all fields NULL had been passed in. Otherwise, if any of the fields are not valid pointers to functions of the correct type, or this parameter is not a valid pointer to a CFArrayCallBacks callbacks structure, the behavior is undefined. If any of the values put into the array is not one understood by one of the callback functions the behavior when that callback function is used is undefined. @result A reference to the new immutable CFArray.