Function coreaudio_sys::CFArrayGetCountOfValue

pub unsafe extern "C" fn CFArrayGetCountOfValue(
    theArray: CFArrayRef, 
    range: CFRange, 
    value: *const c_void
) -> CFIndex

@function CFArrayGetCountOfValue Counts the number of times the given value occurs in the array. @param theArray The array to be searched. If this parameter is not a valid CFArray, the behavior is undefined. @param range The range within the array to search. If the range location or end point (defined by the location plus length minus 1) is outside the index space of the array (0 to N-1 inclusive, where N is the count of the array), the behavior is undefined. If the range length is negative, the behavior is undefined. The range may be empty (length 0). @param value The value for which to find matches in the array. The equal() callback provided when the array was created is used to compare. If the equal() callback was NULL, pointer equality (in C, ==) is used. If value, or any of the values in the array, are not understood by the equal() callback, the behavior is undefined. @result The number of times the given value occurs in the array, within the specified range.