Function coreaudio_sys::CFDictionaryApplyFunction

pub unsafe extern "C" fn CFDictionaryApplyFunction(
    theDict: CFDictionaryRef, 
    applier: CFDictionaryApplierFunction, 
    context: *mut c_void
)

@function CFDictionaryApplyFunction Calls a function once for each value in the dictionary. @param theDict The dictionary to be queried. If this parameter is not a valid CFDictionary, the behavior is undefined. @param applier The callback function to call once for each value in the dictionary. If this parameter is not a pointer to a function of the correct prototype, the behavior is undefined. If there are keys or values which the applier function does not expect or cannot properly apply to, the behavior is undefined. @param context A pointer-sized user-defined value, which is passed as the third parameter to the applier function, but is otherwise unused by this function. If the context is not what is expected by the applier function, the behavior is undefined.