Function coreaudio_sys::CFSetGetValueIfPresent

pub unsafe extern "C" fn CFSetGetValueIfPresent(
    theSet: CFSetRef, 
    candidate: *const c_void, 
    value: *mut *const c_void
) -> Boolean

@function CFSetGetValueIfPresent Retrieves a value in the set which hashes the same as the specified value, if present. @param theSet The set to be queried. If this parameter is not a valid CFSet, the behavior is undefined. @param candidate This value is hashed and compared with values in the set to determine which value to retrieve. The equal() callback provided when the set was created is used to compare. If the equal() callback was NULL, pointer equality (in C, ==) is used. If a value, or any of the values in the set, are not understood by the equal() callback, the behavior is undefined. @param value A pointer to memory which should be filled with the pointer-sized value if a matching value is found. If no match is found, the contents of the storage pointed to by this parameter are undefined. This parameter may be NULL, in which case the value from the dictionary is not returned (but the return value of this function still indicates whether or not the value was present). @result True if the value was present in the set, otherwise false.