Function coreaudio_sys::CFStringTokenizerAdvanceToNextToken

pub unsafe extern "C" fn CFStringTokenizerAdvanceToNextToken(
    tokenizer: CFStringTokenizerRef
) -> CFStringTokenizerTokenType

@function CFStringTokenizerAdvanceToNextToken @abstract Token enumerator. @param tokenizer The reference to CFStringTokenizer returned by CFStringTokenizerCreate. @result Type of the token if succeeded in finding a token and setting it as current token. kCFStringTokenizerTokenNone if failed in finding a token. @discussion If there is no preceding call to CFStringTokenizerGoToTokenAtIndex or CFStringTokenizerAdvanceToNextToken, it finds the first token in the range specified to CFStringTokenizerCreate. If there is a current token after successful call to CFStringTokenizerGoToTokenAtIndex or CFStringTokenizerAdvanceToNextToken, it proceeds to the next token. If succeeded in finding a token, set it as current token and return its token type. Otherwise invalidate current token and return kCFStringTokenizerTokenNone. The range and attribute of the token can be obtained by calling CFStringTokenizerGetCurrentTokenRange and CFStringTokenizerCopyCurrentTokenAttribute. If the token is a compound (with type kCFStringTokenizerTokenHasSubTokensMask or kCFStringTokenizerTokenHasDerivedSubTokensMask), its subtokens and (or) derived subtokens can be obtained by calling CFStringTokenizerGetCurrentSubTokens.