Function coreaudio_sys::MusicDeviceMIDIEvent

pub unsafe extern "C" fn MusicDeviceMIDIEvent(
    inUnit: MusicDeviceComponent, 
    inStatus: UInt32, 
    inData1: UInt32, 
    inData2: UInt32, 
    inOffsetSampleFrame: UInt32
) -> OSStatus

@function MusicDeviceMIDIEvent @abstract Used to sent MIDI channel messages to an audio unit

@discussion This is the API used to send MIDI channel messages to an audio unit. The status and data parameters are used exactly as described by the MIDI specification, including the combination of channel and command in the status byte.

@param inUnit The audio unit @param inStatus The MIDI status byte @param inData1 The first MIDI data byte (value is in the range 0 < 128) @param inData2 The second MIDI data byte (value is in the range 0 < 128). If the MIDI status byte only has one data byte, this should be set to zero. @param inOffsetSampleFrame If you are scheduling the MIDI Event from the audio unit's render thread, then you can supply a sample offset that the audio unit may apply when applying that event in its next audio unit render. This allows you to schedule to the sample, the time when a MIDI command is applied and is particularly important when starting new notes. If you are not scheduling in the audio unit's render thread, then you should set this value to 0

@result noErr, or an audio unit error code