Struct coreaudio_sys::AUParameterMIDIMapping

#[repr(C)]

pub struct AUParameterMIDIMapping {
    pub mScope: AudioUnitScope,
    pub mElement: AudioUnitElement,
    pub mParameterID: AudioUnitParameterID,
    pub mFlags: AUParameterMIDIMappingFlags,
    pub mSubRangeMin: AudioUnitParameterValue,
    pub mSubRangeMax: AudioUnitParameterValue,
    pub mStatus: UInt8,
    pub mData1: UInt8,
    pub reserved1: UInt8,
    pub reserved2: UInt8,
    pub reserved3: UInt32,
}

@struct AUParameterMIDIMapping @abstract Represents a mapping between a MIDI message and an audio unit's parameter. @discussion The reserved fields in this structure are for future use. In the current implementation, they help align the structure to 64 bit size. Do not use the names of these fields in a host application. They are subject to change.

Fields
mScope: AudioUnitScope mElement: AudioUnitElement mParameterID: AudioUnitParameterID mFlags: AUParameterMIDIMappingFlags mSubRangeMin: AudioUnitParameterValue mSubRangeMax: AudioUnitParameterValue mStatus: UInt8 mData1: UInt8 reserved1: UInt8 reserved2: UInt8 reserved3: UInt32