Type Definition coreaudio_sys::AUParameterMIDIMappingFlags

type AUParameterMIDIMappingFlags = UInt32;

@enum ParameterMIDIMappings @abstract General defined values to customize the behavior of parameter-to-MIDI mappings

@constant kAUParameterMIDIMapping_AnyChannelFlag @discussion If this flag is set and the value of the mStatus field is a MIDI channel message, then the MIDI channel number in the status byte is ignored; the mapping is from the specified MIDI message on any channel.

@constant kAUParameterMIDIMapping_AnyNoteFlag @discussion If this flag is set and the value of the mStatus field is a Note On, Note Off, or Polyphonic Pressure message, the message's note number is ignored. The mapping is from any note number.

@constant kAUParameterMIDIMapping_SubRange @discussion Set this flag if the MIDI control should map only to a sub-range of the parameter's value. Then specify that range in the mSubRangeMin and mSubRangeMax member fields.

@constant kAUParameterMIDIMapping_Toggle @discussion Intended for Boolean typed parameters. When this property is set, it means that the parameter's value should be toggled when the mapped MIDI message is received. For example, if the parameter's value is currently TRUE, when the mapped MIDI message is received the value changes to FALSE.

@constant kAUParameterMIDIMapping_Bipolar @discussion This property is useful when mapping a parameter to a MIDI Controller. When set, it indicates that the parameter can assume only two values: on or off. For this reason, a parameter associated with this property is typically Boolean. For example, if this property is set for a parameter mapped to a sustain pedal MIDI controller, controller values from 0 to 64 result in the parameter switched to its "off" state; controller values from 65 to 127 result in the parameter switched to its "on" state.

This property works in connection with the kAUParameterMIDIMapping_Bipolar_On property. The value of the kAUParameterMIDIMapping_Bipolar_On property

@constant kAUParameterMIDIMapping_Bipolar_On @discussion Determines whether the "on" state of a parameter is mapped to the "on" or "off" state of the associated MIDI controller. Only valid if the kAUParameterMIDIMapping_Bipolar property is set.