Type Definition coreaudio_sys::AudioComponent

type AudioComponent = *mut OpaqueAudioComponent;

@typedef AudioComponent @abstract The type used to represent a class of particular audio components @discussion An audio component is usually found through a search and is then uniquely identified by the triple of an audio component's type, subtype and manufacturer.

It can have properties associated with it (such as a name, a version).

It is then used as a factory (like a class in an object-oriented programming language) from which to create instances. The instances are used to do the actual work.

For example: the AudioComponentDescription 'aufx'/'dely'/'appl' describes the delay audio unit effect from Apple, Inc. You can find this component by searching explicitly for the audio component that matches this pattern (this is an unique identifier - there is only one match to this triple ID). Then once found, instances of the Apple delay effect audio unit can be created from its audio component and used to apply that effect to an audio signal. A single component can create any number of component instances.