Type Definition coreaudio_sys::CFDictionaryRetainCallBack

type CFDictionaryRetainCallBack = Option<unsafe extern "C" fn(allocator: CFAllocatorRef, value: *const c_void) -> *const c_void>;

@typedef CFDictionaryKeyCallBacks Structure containing the callbacks for keys of a CFDictionary. @field version The version number of the structure type being passed in as a parameter to the CFDictionary creation functions. This structure is version 0. @field retain The callback used to add a retain for the dictionary on keys as they are used to put values into the dictionary. This callback returns the value to use as the key in the dictionary, which is usually the value parameter passed to this callback, but may be a different value if a different value should be used as the key. The dictionary's allocator is passed as the first argument. @field release The callback used to remove a retain previously added for the dictionary from keys as their values are removed from the dictionary. The dictionary's allocator is passed as the first argument. @field copyDescription The callback used to create a descriptive string representation of each key in the dictionary. This is used by the CFCopyDescription() function. @field equal The callback used to compare keys in the dictionary for equality. @field hash The callback used to compute a hash code for keys as they are used to access, add, or remove values in the dictionary.